feat: complete WebGL 2.0 Fast3D renderer, 2P co-op & 16P multiplayer netplay

+++ src/game/bondinv.c
@@ -1195,7 +1195,7 @@ f32 bondinvGetHoffsetForIndex(int index)
     return get_horizontal_offset_on_solo_watch_menu_for_item(bondinvGetTextbyInvIndex(index));
 }
 
-int bondinvGetVoffsetForIndex(int index)
+f32 bondinvGetVoffsetForIndex(int index)
 {
     return get_vertical_offset_on_solo_watch_menu_for_item(bondinvGetTextbyInvIndex(index));
 }
@@ -1330,12 +1330,12 @@ f32 bondinvGetDifferent45AngleForIndex(int index)
     return get_45_degree_angle(bondinvGetTextbyInvIndex(index));
 }
 
-int bondinvGetVposWatchForIndex(int index)
+f32 bondinvGetVposWatchForIndex(int index)
 {
     return get_vertical_position_solo_watch_menu_main_page_for_item(bondinvGetTextbyInvIndex(index));
 }
 
-int bondinvGetHposWatchForIndex(int index)
+f32 bondinvGetHposWatchForIndex(int index)
 {
     return get_lateral_position_solo_watch_menu_main_page_for_item(bondinvGetTextbyInvIndex(index));
 }
@@ -1345,12 +1345,12 @@ f32 bondinvGetDepthWatchForIndex(int index)
     return get_depth_on_solo_watch_menu_page_for_item(bondinvGetTextbyInvIndex(index));
 }
 
-int bondinvGetXrotWatchForIndex(int index)
+f32 bondinvGetXrotWatchForIndex(int index)
 {
     return get_xrotation_solo_watch_menu_for_item(bondinvGetTextbyInvIndex(index));
 }
 
-int bondinvGetYrotWatchForIndex(int index)
+f32 bondinvGetYrotWatchForIndex(int index)
 {
     return get_yrotation_solo_watch_menu_for_item(bondinvGetTextbyInvIndex(index));
 }
